@@ -2,9 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\File;
-use App\FitAnalyser;
+use App\Models\File;
+use App\Models\FitAnalyser;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FileController extends Controller
 {
@@ -47,22 +48,24 @@ class FileController extends Controller
      */
     public function show(File $file)
     {
-        // This should probably be extracted to its own method
-        // as it's not really the responsibility of show
-        $path = storage_path('app/' . $file->investigation->id . '/' . $file->name);
-        $fitData = new FitAnalyser($path);
-        if ( ! $file->parsed) {
-            $file->type = $fitData->getType();
-            $file->parsed = true;
-            $file->save();
-        }
+        if (Auth::user()->can('access', $file)) {
+            // This should probably be extracted to its own method
+            // as it's not really the responsibility of show
+            $path = storage_path('app/' . $file->investigation->id . '/' . $file->name);
+            $fitData = new FitAnalyser($path);
+            if ( ! $file->parsed) {
+                $file->type = $fitData->getType();
+                $file->parsed = true;
+                $file->save();
+            }
 
-//        var_dump($fitData->data_mesgs);
+//        var_dump($path);
 
-        $currentHash = hash('sha256', file_get_contents($path));
+            $currentHash = hash('sha256', file_get_contents($path));
 //        $json = json_encode($fitData->data_mesgs);
 
-        return view('file.show', ['file' => $file, 'data' => $fitData, 'currentHash' => $currentHash]);
+            return view('files.show', ['file' => $file, 'data' => $fitData, 'currentHash' => $currentHash]);
+        }
     }
 
     /**
