@@ -1,44 +1,48 @@
-@extends('layouts.app')
+<x-guest-layout>
+    <x-jet-authentication-card>
+        <x-slot name="logo">
+            <x-jet-authentication-card-logo />
+        </x-slot>
 
-@section('content')
-<div>
-    <h1>{{ __('Login') }}</h1>
+        <x-jet-validation-errors class="mb-4" />
 
-    <form method="POST" action="{{ route('login') }}">
-        @csrf
-
-        <label for="email">{{ __('E-Mail Address') }}</label>
-        <input id="email" type="email" class="@error('email') some-error-class @enderror" name="email" value="{{ old('email') }}" required autocomplete="email" autofocus>
-
-        @error('email')
-            <span role="alert">
-                <strong>{{ $message }}</strong>
-            </span>
-        @enderror
-
-        <label for="password">{{ __('Password') }}</label>
-        <input id="password" type="password" class="@error('password') some-error-class @enderror" name="password" required autocomplete="current-password">
-
-        @error('password')
-            <span role="alert">
-                <strong>{{ $message }}</strong>
-            </span>
-        @enderror
-
-        <input type="checkbox" name="remember" id="remember" {{ old('remember') ? 'checked' : '' }}>
-        <label for="remember">
-            {{ __('Remember Me') }}
-        </label>
-
-        <button type="submit">
-            {{ __('Login') }}
-        </button>
-
-        @if (Route::has('password.request'))
-            <a href="{{ route('password.request') }}">
-                {{ __('Forgot Your Password?') }}
-            </a>
+        @if (session('status'))
+            <div class="mb-4 font-medium text-sm text-green-600">
+                {{ session('status') }}
+            </div>
         @endif
-    </form>
-</div>
-@endsection
+
+        <form method="POST" action="{{ route('login') }}">
+            @csrf
+
+            <div>
+                <x-jet-label for="email" value="{{ __('Email') }}" />
+                <x-jet-input id="email" class="block mt-1 w-full" type="email" name="email" :value="old('email')" required autofocus />
+            </div>
+
+            <div class="mt-4">
+                <x-jet-label for="password" value="{{ __('Password') }}" />
+                <x-jet-input id="password" class="block mt-1 w-full" type="password" name="password" required autocomplete="current-password" />
+            </div>
+
+            <div class="block mt-4">
+                <label for="remember_me" class="flex items-center">
+                    <input id="remember_me" type="checkbox" class="form-checkbox" name="remember">
+                    <span class="ml-2 text-sm text-gray-600">{{ __('Remember me') }}</span>
+                </label>
+            </div>
+
+            <div class="flex items-center justify-end mt-4">
+                @if (Route::has('password.request'))
+                    <a class="underline text-sm text-gray-600 hover:text-gray-900" href="{{ route('password.request') }}">
+                        {{ __('Forgot your password?') }}
+                    </a>
+                @endif
+
+                <x-jet-button class="ml-4">
+                    {{ __('Login') }}
+                </x-jet-button>
+            </div>
+        </form>
+    </x-jet-authentication-card>
+</x-guest-layout>
