@@ -1,45 +1,43 @@
-@extends('layouts.app')
+<x-guest-layout>
+    <x-jet-authentication-card>
+        <x-slot name="logo">
+            <x-jet-authentication-card-logo />
+        </x-slot>
 
-@section('content')
-<div>
-    <h1>{{ __('Register') }}</h1>
+        <x-jet-validation-errors class="mb-4" />
 
-    <form method="POST" action="{{ route('register') }}">
-        @csrf
+        <form method="POST" action="{{ route('register') }}">
+            @csrf
 
-        <label for="name">{{ __('Name') }}</label>
-        <input id="name" type="text" class="@error('name') some-error-class @enderror" name="name" value="{{ old('name') }}" required autocomplete="name" autofocus>
+            <div>
+                <x-jet-label for="name" value="{{ __('Name') }}" />
+                <x-jet-input id="name" class="block mt-1 w-full" type="text" name="name" :value="old('name')" required autofocus autocomplete="name" />
+            </div>
 
-        @error('name')
-            <span role="alert">
-                <strong>{{ $message }}</strong>
-            </span>
-        @enderror
+            <div class="mt-4">
+                <x-jet-label for="email" value="{{ __('Email') }}" />
+                <x-jet-input id="email" class="block mt-1 w-full" type="email" name="email" :value="old('email')" required />
+            </div>
 
-        <label for="email">{{ __('E-Mail Address') }}</label>
-        <input id="email" type="email" class="@error('email') some-error-class @enderror" name="email" value="{{ old('email') }}" required autocomplete="email">
+            <div class="mt-4">
+                <x-jet-label for="password" value="{{ __('Password') }}" />
+                <x-jet-input id="password" class="block mt-1 w-full" type="password" name="password" required autocomplete="new-password" />
+            </div>
 
-        @error('email')
-            <span role="alert">
-                <strong>{{ $message }}</strong>
-            </span>
-        @enderror
+            <div class="mt-4">
+                <x-jet-label for="password_confirmation" value="{{ __('Confirm Password') }}" />
+                <x-jet-input id="password_confirmation" class="block mt-1 w-full" type="password" name="password_confirmation" required autocomplete="new-password" />
+            </div>
 
-        <label for="password">{{ __('Password') }}</label>
-        <input id="password" type="password" class="@error('password') some-error-class @enderror" name="password" required autocomplete="new-password">
+            <div class="flex items-center justify-end mt-4">
+                <a class="underline text-sm text-gray-600 hover:text-gray-900" href="{{ route('login') }}">
+                    {{ __('Already registered?') }}
+                </a>
 
-        @error('password')
-            <span role="alert">
-                <strong>{{ $message }}</strong>
-            </span>
-        @enderror
-
-        <label for="password-confirm">{{ __('Confirm Password') }}</label>
-        <input id="password-confirm" type="password" name="password_confirmation" required autocomplete="new-password">
-
-        <button type="submit">
-            {{ __('Register') }}
-        </button>
-    </form>
-</div>
-@endsection
+                <x-jet-button class="ml-4">
+                    {{ __('Register') }}
+                </x-jet-button>
+            </div>
+        </form>
+    </x-jet-authentication-card>
+</x-guest-layout>
