@@ -1,6 +1,4 @@
 const mix = require('laravel-mix');
-const domain = 'fit-forensics.test';
-const homedir = require('os').homedir();
 
 /*
  |--------------------------------------------------------------------------
@@ -8,23 +6,13 @@ const homedir = require('os').homedir();
  |--------------------------------------------------------------------------
  |
  | Mix provides a clean, fluent API for defining some Webpack build steps
- | for your Laravel application. By default, we are compiling the Sass
+ | for your Laravel applications. By default, we are compiling the CSS
  | file for the application as well as bundling up all the JS files.
  |
  */
 
-mix.postCss('resources/css/app.css', 'public/css', [
-    require('tailwindcss'),
-    require('autoprefixer'),
-])
-    .browserSync({
-        files: ['resources/views/**/*.php'],
-        proxy: 'https://' + domain,
-        host: domain,
-        open: 'external',
-        notify: false,
-        https: {
-            key: homedir + '/.config/valet/Certificates/' + domain + '.key',
-            cert: homedir + '/.config/valet/Certificates/' + domain + '.crt',
-        },
-    });
+mix.js('resources/js/app.js', 'public/js')
+    .postCss('resources/css/app.css', 'public/css', [
+        require('postcss-import'),
+        require('tailwindcss'),
+    ]);
